add some test_case
